pyproject.toml
[project]
name = "polymarket-skill"
version = "0.1.0"
description = "OpenClaw skill wrapper for Polymarket CLI"
readme = "SKILL.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]
